[decoder] Add ability to select different instruction decoders
Added the ability to select a logical and table based instruction
decoders. If not provided, it defaults to table based. If the
table based decoder is selected, instructions will be decoded
based on the information generated from generate_isa.py. The
logcial decoder looks at instructions based on x, y and z parts
of the opcode and attempts to determine what the instruction is
based on patterns.